lab10/requirements.txt
asyncio-paho>=0.6.0
bleak>=0.21.1
paho-mqtt>=1.6.1
ruuvitag-sensor>=2.3.0
